chore(InputMasked): rename files to ts/tsx
