+++ packages/dnb-eufemia/src/components/input-masked/InputMaskedHooks.tsx
@@ -1,3 +1,5 @@
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-nocheck
 /**
  * Web InputMasked Component
  *
