+++ packages/dnb-eufemia/src/components/input-masked/style/index.ts
