+++ packages/dnb-eufemia/src/components/input-masked/text-mask/conformToMask.ts
@@ -1,3 +1,5 @@
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-nocheck
 import {
   convertMaskToPlaceholder,
   isArray,
