+++ packages/dnb-eufemia/src/components/input-masked/text-mask/constants.ts
