+++ packages/dnb-eufemia/src/components/input-masked/text-mask/utilities.ts
@@ -1,3 +1,6 @@
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-nocheck
+
 import { placeholderChar as defaultPlaceholderChar } from './constants'
 
 const emptyArray = []
